added time stamps, and hide/delete checked guards against being clicked when nothing is checked

+++ Public/Scripts/results.js
@@ -18,8 +18,11 @@ const resultsJs = (function() {
   function handleHideCheckedClicked() {
     $("#stuff").on("click", ".hideAll", event => {
       event.preventDefault();
-      store.checked = true;
-      eventHandlers.render();
+      let newArray = store.items.filter(item => item.checked === "true");
+      if (newArray.length > 0) {
+        store.checked = true;
+        eventHandlers.render();
+      }
     });
   }
 
@@ -31,48 +34,49 @@ const resultsJs = (function() {
     });
   }
 
-
   function handleItemCheckClicked() {
     $("#stuff").on("click", ".checkbox", event => {
       const id = getItemIdFromElement(event.currentTarget);
       const thisItem = store.items.find(item => item.id === id);
 
-      if (thisItem.checked === 'false') {
-          thisItem.checked = 'true';
+      if (thisItem.checked === "false") {
+        thisItem.checked = "true";
+      } else if (thisItem.checked === "true") {
+        thisItem.checked = "false";
       }
-      else if (thisItem.checked === 'true') {
-          thisItem.checked = 'false';
-      }
-      
 
-      api.update(`/results/${id}`, thisItem)
+      api
+        .update(`/results/${id}`, thisItem)
         .then(() => {
-            eventHandlers.render();
-      })
+          eventHandlers.render();
+        })
         .catch(err => next(err));
     });
   }
 
   function handleDeleteCheckedClicked() {
     $("#stuff").on("click", ".deleteAll", event => {
-      let result = confirm(
-        "Are you sure? This is permanent and cannot be undone."
-      );
-      if (result) {
-        let deleteItems = store.items.filter(item => item.checked === 'true');
-        console.log(deleteItems);
-        for (let i = 0; i < deleteItems.length; i++) {
-          let id = deleteItems[i].id;
-          console.log(id);
-          api
-            .deleteItem(`/results/${id}`)
-            .then(() => {
-              console.log(store.items.filter(item => item.id === id));
-              store.findAndDelete(id);
-            })
-            .then(() => {
-              eventHandlers.render();
-            });
+      let newArray = store.items.filter(item => item.checked === "true");
+      if (newArray.length > 0) {
+        let result = confirm(
+          `Are you sure? This will permanently delete ${newArray.length} items.`
+        );
+        if (result) {
+          let deleteItems = store.items.filter(item => item.checked === "true");
+          console.log(deleteItems);
+          for (let i = 0; i < deleteItems.length; i++) {
+            let id = deleteItems[i].id;
+            console.log(id);
+            api
+              .deleteItem(`/results/${id}`)
+              .then(() => {
+                console.log(store.items.filter(item => item.id === id));
+                store.findAndDelete(id);
+              })
+              .then(() => {
+                eventHandlers.render();
+              });
+          }
         }
       }
     });
